refactor(sessions): rename session manager module
