+++ packages/sessions/src/__tests__/credential-manager.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test, beforeEach } from "bun:test";
 import type { PermissionScopeType } from "@xmtp/signet-schemas";
-import { createCredentialManager } from "../session-manager.js";
-import type { InternalCredentialManager } from "../session-manager.js";
+import { createCredentialManager } from "../credential-manager.js";
+import type { InternalCredentialManager } from "../credential-manager.js";
 import { createTestCredentialConfig, baseScopes } from "./fixtures.js";
 
 let manager: InternalCredentialManager;
