+++ packages/sessions/src/__tests__/update-actions.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, test } from "bun:test";
 import type { PermissionScopeType } from "@xmtp/signet-schemas";
 import type { CredentialManager } from "@xmtp/signet-contracts";
-import { createCredentialManager } from "../session-manager.js";
+import { createCredentialManager } from "../credential-manager.js";
 import { createCredentialService } from "../service.js";
 import { createUpdateActions } from "../update-actions.js";
 import type { UpdateActionDeps } from "../update-actions.js";
@@ -10,7 +10,7 @@ import {
   baseScopes,
   createTestScopes,
 } from "./fixtures.js";
-import type { InternalCredentialManager } from "../session-manager.js";
+import type { InternalCredentialManager } from "../credential-manager.js";
 
 let manager: InternalCredentialManager;
 let credentialService: CredentialManager;
