+++ packages/sessions/src/credential-manager.ts
